requirements.txt
pdfplumber
pandas
python-docx

absl-py==2.2.2
adal==1.2.7
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.0.0
applicationinsights==0.11.10
argcomplete==3.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==22.2.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.32.0
azure-graphrbac==0.61.2
azure-identity==1.21.0
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.5.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-network==28.1.0
azure-mgmt-resource==23.3.0
azure-mgmt-storage==22.0.0
azureml-automl-core==1.59.0
azureml-core==1.59.0.post1
azureml-dataprep==5.1.6
azureml-dataprep-native==41.0.0
azureml-dataprep-rslex==2.22.5
azureml-dataset-runtime==1.59.0
azureml-pipeline==1.59.0
azureml-pipeline-core==1.59.0
azureml-pipeline-steps==1.59.0
azureml-sdk==1.59.0
azureml-telemetry==1.59.0
azureml-train-automl-client==1.59.0
azureml-train-core==1.59.0
azureml-train-restclients-hyperdrive==1.59.0
Babel==2.13.1
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.3.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.8
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.0
contextlib2==21.6.0
contourpy==1.3.2
cryptography==44.0.2
ctgan==0.11.0
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==2.1
docker==7.1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==2.0.1
faiss-cpu==1.10.0
Faker==37.1.0
fastapi==0.115.12
fastjsonschema==2.19.0
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.58.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
fusepy==3.0.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.2.2
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
idna==3.4
importlib_metadata==8.2.0
importlib_resources==6.4.0
iniconfig==1.1.1
ipykernel==6.27.0
ipython==8.17.2
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.9.0
Jinja2==3.1.2
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
json5==0.9.14
jsonpatch==1.33
jsonpickle==4.0.2
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.2
keras==3.9.2
kiwisolver==1.4.8
knack==0.12.0
langchain==0.0.200
langchain-community==0.3.24
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langchainplus-sdk==0.0.20
langsmith==0.3.42
libclang==18.1.1
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml_dtypes==0.5.1
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.4.3
mypy_extensions==1.1.0
namex==0.0.9
narwhals==1.27.1
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
ndg-httpsclient==0.5.1
nest-asyncio==1.5.8
networkx==3.4.2
notebook==7.0.6
notebook_shim==0.2.3
numexpr==2.10.2
numpy==1.23.5
oauthlib==3.2.2
openai==1.79.0
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.18
overrides==7.4.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
paramiko==3.5.1
parso==0.8.3
pathspec==0.12.1
pickle-mixin==1.0.2
pillow==11.1.0
pkginfo==1.12.1.2
platformdirs==4.0.0
pluggy==1.0.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
propcache==0.3.1
protobuf==5.29.3
psutil==5.9.6
pure-eval==0.2.2
pyarrow==19.0.0
pyasn1==0.6.1
pycparser==2.21
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.17.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pyparsing==3.2.3
pypdf==5.5.0
PyPDF2==3.0.1
pyreadline3==3.5.4
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.1.0
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
rdt==1.16.0
referencing==0.31.0
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.13.1
scikit-learn==1.6.1
scipy==1.15.2
SecretStorage==3.3.3
Send2Trash==1.8.2
shell==1.0.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
streamlit==1.42.1
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
terminado==0.18.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==2.7.0
tornado==6.3.3
tqdm==4.67.1
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.0.3
uvicorn==0.34.2
watchdog==6.0.0
wcwidth==0.2.11
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
